PageWrapper component POC
Proof Of Concept, of how to segment the transcript into 'pages' or arbitrary time length eg 1 hour interval. Ideally this would be brought inside the component, rather then on the outside. So that things like export and save  work across all of it, but things like sync only at the page level